Create cliphist-rofi-img-strip-id
Allows you to use ```cliphist-rofi-img``` without seeing the id prefix in each entry. Since the ```display-columns``` flag only works with the ```dmenu``` mode, it will not work with the aforementioned script. This script works off of that script, but manually strips the id when passing it off to rofi and reconstructs rofi's output to the original string.